+++ apps/react-namespace/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lodado/react-namespace
 
+## 1.0.2
+
+### Patch Changes
+
+- move promise handler from namespace to react-namepspace repository
+- Updated dependencies
+  - @lodado/namespace-core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes